% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixRFb.r
\name{MixRFb}
\alias{MixRFb}
\title{Mixed Logistic Random Forest for Binary Data}
\usage{
MixRFb(Y, x, random, data, initialRandomEffects = 0, ErrorTolerance = 0.001,
  MaxIterations = 200, ErrorTolerance0 = 0.001, MaxIterations0 = 15,
  verbose = FALSE)
}
\arguments{
\item{Y}{The outcome variable.}

\item{x}{A formula string contains the predictors.}

\item{random}{A string in lme4 format indicates the random effect model.}

\item{data}{The data set as a data frame.}

\item{initialRandomEffects}{The initial values for random effects.}

\item{ErrorTolerance}{The tolerance for log-likelihood.}

\item{MaxIterations}{The maximum iteration times for each run of PQL.}

\item{ErrorTolerance0}{The tolerance for eta (penalized quasi-likelihood, PQL).}

\item{MaxIterations0}{The maximum iteration times for PQL.}

\item{verbose}{The option to monitor each run of PQL or not.}
}
\value{
A list contains the random forest, mixed model, and random effects.
See the example below for the usage. A predict() function is also available below.
}
\description{
Mixed Logistic Random Forest for Binary Data
}
\examples{

# example data (http://stats.stackexchange.com/questions/70783/how-to-assess-the-fit-of-a-binomial-glmm-fitted-with-lme4-1-0)
dat <- read.table("http://pastebin.com/raw.php?i=vRy66Bif")

library(party)
library(lme4)

source('MixRFb.r')
system.time(tmp <- MixRFb(Y=dat$true, x='factor(distance) + consequent + factor(direction) + factor(dist)', random='(1|V1)',
                          data=dat, initialRandomEffects=0,
                          ErrorTolerance=1, MaxIterations=200,
                          ErrorTolerance0=0.3, MaxIterations0=15, verbose=T))

# tmp$forest
# tmp$MixedModel
# tmp$RandomEffects

# eta
pred1 = predict.MixRF(tmp, dat, EstimateRE=TRUE)
prob = 1/(1+exp(-pred1))
res = (prob>.5)

# classification
table(res,dat$true)
}

