% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.MixRF.r
\name{predict.MixRF}
\alias{predict.MixRF}
\title{Prediction Function for MixRF}
\usage{
\method{predict}{MixRF}(object, newdata, id = NULL, EstimateRE = TRUE)
}
\arguments{
\item{object}{The fitted MixRF object.}

\item{newdata}{A data frame contains the predictors for prediction.}

\item{id}{The group variable in the new data.}

\item{EstimateRE}{To use the estimated random effects in the prediction or not. The default is TRUE.}
}
\value{
A matrix (now for balanced data) contains the predicted values.
}
\description{
Prediction Function for MixRF
}
\examples{


library(lme4)
library(randomForest)
data(sleepstudy)

tmp = MixRF(Y=sleepstudy$Reaction, x=as.data.frame(sleepstudy$Days), random='(Days|Subject)',
            data=sleepstudy, initialRandomEffects=0, ErrorTolerance=0.01, MaxIterations=100)

pred = predict.MixRF(object=tmp, newdata=sleepstudy, EstimateRE=TRUE)
}

